requirements.txt
playwright
pytest
pytest-xdist
allure-pytest
playwright-stealth
